protectors: Add TPM2 Key Protector
The TPM2 key protector is a module that enables the automatic retrieval
of a fully-encrypted disk's unlocking key from a TPM 2.0.

The theory of operation is such that the module accepts various
arguments, most of which are optional and therefore possess reasonable
defaults. One of these arguments is the keyfile parameter, which is
mandatory. There are two supported key formats:

1. Raw Sealed Key (--keyfile)
   The raw sealed key glues TPM2B_PUBLIC and TPM2B_PRIVATE, which are
   returned from TPM2_Create, into one file, and is defined as a C
   struct:

   typedef struct TPM2_SEALED_KEY {
     TPM2B_PUBLIC  public;
     TPM2B_PRIVATE private;
   } TPM2_SEALED_KEY;

2. TPM 2.0 Key (--tpm2key)
   The following is the ASN.1 definition of TPM 2.0 Key File:

   TPMKey ::= SEQUENCE {
    type        OBJECT IDENTIFIER
    emptyAuth   [0] EXPLICIT BOOLEAN OPTIONAL
    policy      [1] EXPLICIT SEQUENCE OF TPMPolicy OPTIONAL
    secret      [2] EXPLICIT OCTET STRING OPTIONAL
    authPolicy  [3] EXPLICIT SEQUENCE OF TPMAuthPolicy OPTIONAL
    parent      INTEGER
    pubkey      OCTET STRING
    privkey     OCTET STRING
  }

  The TPM2 key protector only expects a "sealed" key in DER encoding,
  so 'type' is always 2.23.133.10.1.5, 'emptyAuth' is 'TRUE', and
  'secret' is empty. TPM2B_PUBLIC is stored in 'pubkey' and
  TPM2B_PRIVATE is stored in 'privkey'. 'policy' and 'authPolicy' are
  ignored for the time being and will be used for the advanced features
  in the future.

  For more details: https://www.hansenpartnership.com/draft-bottomley-tpm2-keys.html

This sealed key file is created via the grub-protect tool. The tool
utilizes the TPM's sealing functionality to seal (i.e., encrypt) an
unlocking key using a Storage Root Key (SRK) to the values of various
Platform Configuration Registers (PCRs). These PCRs reflect the state
of the system as it boots. If the values are as expected, the system
may be considered trustworthy, at which point the TPM allows for a
caller to utilize the private component of the SRK to unseal (i.e.,
decrypt) the sealed key file. The caller, in this case, is this key
protector.

The TPM2 key protector registers two commands:

- tpm2_key_protector_init: Initializes the state of the TPM2 key
                           protector for later usage, clearing any
                           previous state, too, if any.

- tpm2_key_protector_clear: Clears any state set by tpm2_key_protector_init.

The way this is expected to be used requires the user to, either
interactively or, normally, via a boot script, initialize (i.e.,
configure) the key protector and then specify that it be used by the
'cryptomount' command (modifications to this command are in a different
patch).

For instance, to unseal the raw sealed key file:

tpm2_key_protector_init --keyfile=(hd0,gpt1)/efi/grub2/sealed-1.key
cryptomount DISK1 -P tpm2

tpm2_key_protector_init --keyfile=(hd0,gpt1)/efi/grub2/sealed-2.key --pcrs=7,11
cryptomount DISK2 -P tpm2

Or, to unseal the TPM 2.0 Key file:

tpm2_key_protector_init --tpm2key=(hd0,gpt1)/efi/grub2/sealed-1.tpm
cryptomount DISK1 -P tpm2

tpm2_key_protector_init --tpm2key=(hd0,gpt1)/efi/grub2/sealed-2.tpm --pcrs=7,11
cryptomount DISK2 -P tpm2

If a user does not initialize the key protector and attempts to use it anyway,
the protector returns an error.

Signed-off-by: Hernan Gatta <hegatta@linux.microsoft.com>
Signed-off-by: Gary Lin <glin@suse.com>