tpm2: declare the input arguments of TPM2 functions as const
The arguments, except the buffer, of the marshal functions are the pure
inputs. Also, the TPM2 command parameters are supposed not changed by
the command. Declare those arguments as 'const' so that the compiler can
help to detect the undesired change on those arguments. Besides, when
looking up the TPM2 command functions, it's easier to tell which arguments
are the command parameters (the constant arguments) and which are the
response parameters.

Signed-off-by: Gary Lin <glin@suse.com>