tpm2: Enable tpm2 module for grub-emu
As a preparation to test TPM 2.0 TSS stack with grub-emu, the new
option, --tpm-device, is introduced to specify the TPM device for
grub-emu so that grub-emu can share the emulated TPM device with
the host.

Since grub-emu can directly access the device node on host, it's easy to
implement the essential TCG2 command with the standard read/write
functions and enable the commands from tpm2 module for grub-emu, so that
we can further test TPM key unsealing with grub-emu.

Signed-off-by: Gary Lin <glin@suse.com>