tpm2_key_protector: Unseal key from a buffer
As the preparation to load the sealed key from the NV index handle,
the logic to handle the file buffer is extracted as an independent
function and the SRK recover function only reads the file and sends the
file buffer to the new function. Besides, the file format is detected
automatically before unmarshalling the data, so there is no need to use
the command option to specify the file format anymore. In other words,
'--tpm2key' and '--keyfile' are the same now.

Signed-off-by: Gary Lin <glin@suse.com>