util: Introduce grub-mkpasswd-argon2
This commit adds grub-mkpasswd-argon2, a new user-space utility used to
generate Argon2 password hashes compatible with the password_argon2
command.

The utility generates a random salt and computes the hash based on the
user's input password. It allows customizing the Argon2 parameters via
command-line arguments.

The default parameters are configured as follows:
- Iterations (-i): 5
- Memory (-m): 131072 KiB
- Parallelism (-p): 1
- Salt length (-s): 16 bytes
- Hash length (-l): 32 bytes

Signed-off-by: Gary Lin <glin@suse.com>